ref: isolate sort_on_flexibility function to flexibility module
